navedex/requirements.txt
asgiref==3.2.10
Django==3.0.8
django-filter==2.3.0
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
factory-boy==3.0.1
Faker==4.1.1
psycopg2==2.8.5
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
text-unidecode==1.3
